Moved classes to the same package
