+++ cotgenerator/src/main/AndroidManifest.xml
@@ -29,7 +29,7 @@
         <activity android:name="com.jon.common.ui.ListPresetsActivity" />
 
         <service
-            android:name="com.jon.cotgenerator.service.GeneratorService"
+            android:name="com.jon.cotgenerator.GeneratorService"
             android:enabled="true"
             android:exported="false" />
