+++ cotgenerator/src/main/res/values/pref_strings.xml
@@ -3,6 +3,7 @@
     <string name="randomCallsigns">Use Random Icon Callsigns?</string>
     <string name="randomColour">Use Random Team Colours?</string>
     <string name="randomRole">Use Random Team Roles?</string>
+    <string name="iconCount">Icon Count</string>
 
     <string name="locationSettings">Location Settings</string>
     <string name="followMyGpsLocation">Follow My GPS Location?</string>
