+++ cotgenerator/src/main/res/xml/settings.xml
@@ -203,8 +203,4 @@
 
     </PreferenceCategory>
 
-    <EditTextPreference
-        android:key="newPresetAdded"
-        app:defaultValue="0" />
-
 </PreferenceScreen>
